ARROWGRAM::PARSE-ON-PATHS to returns a list of results
Improve the PARSE-ON-PATHS macro to return the values returned by
executing its body as a list.

Split the tests into two files.  The tests in 'basic.lisp' assume that
the ARROWGRAM::PARSE-ON-PATHS macro is working; the test in 'dep.lisp'
works out Paul's version of the main loop in detail.